python-backend/requirements.txt
attrs==22.2.0
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
constantly
cryptography
cssselect
docker<=6.0.1
filelock==3.9.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath<1.0.0
lxml==4.9.2
packaging==23.0
parsel==1.7.0
Protego==0.2.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
PyDispatcher==2.0.7
pyOpenSSL==23.0.0
PyYAML<5.4
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
retrying==1.3.4
scrapinghub==2.4.0
Scrapy==2.8.0
service-identity==21.1.0
shub==2.14.4
six==1.16.0
SQLAlchemy==2.0.4
tldextract==3.4.0
toml==0.10.2
tqdm==4.55.1
Twisted==22.10.0
typing_extensions==4.5.0
urllib3<1.26
w3lib==2.1.1
websocket-client<0.55.0
zope.interface==5.5.2
